Separate parameter functions from data functions for GGSW and GLWE
